requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.6.0
beautifulsoup4==4.12.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
crispy-bootstrap4==2022.1
cryptography==40.0.2
defusedxml==0.7.1
Django==4.2
django-colorfield==0.8.0
django-crispy-forms==2.0
django-environ==0.10.0
django-recaptcha==3.0.0
django-tinymce==3.6.1
idna==3.4
oauthlib==3.2.2
Pillow==9.5.0
psycopg2==2.9.6
pycparser==2.21
PyJWT==2.6.0
python-environ==0.4.54
python3-openid==3.2.0
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
urllib3==1.26.15
